// src/tsco-cli/enums/module-member-type.ts
export enum ModuleMemberType
{
    enums = "enums",
    exportedEnums = "exportedEnums",
    types = "types",
    exportedTypes = "exportedTypes",
    interfaces = "interfaces",
    exportedInterfaces = "exportedInterfaces",
    classes = "classes",
    exportedClasses = "exportedClasses",
    functions = "functions",
    exportedFunctions = "exportedFunctions",
    constants = "constants",
    variables = "variables",
    exportedConstants = "exportedConstants",
    exportedVariables = "exportedVariables"
}